wip: clean up main file a bit
