+++ src/types.ts
@@ -1,0 +1,2 @@
+export type Point = [number, number];
+export type Polyline = Point[];
